requirements.txt.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
supabase
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart